LaboEntity fixed

+++ gestioncours/spring-server/src/main/java/ch/heigvd/amt/gestioncours/entities/LaboEntity.java
@@ -13,7 +13,7 @@ public class LaboEntity implements Serializable {
     @Id
     @Column(name= "laboName", unique= true, nullable=false)
     private String laboName;
-    private Long ponderation;
+    private Integer ponderation;
 
     public String getLaboName() {
         return laboName;
@@ -23,11 +23,11 @@ public class LaboEntity implements Serializable {
         this.laboName = laboName;
     }
 
-    public Long getPonderation() {
+    public Integer getPonderation() {
         return ponderation;
     }
 
-    public void setPonderation(Long ponderation) {
+    public void setPonderation(Integer ponderation) {
         this.ponderation = ponderation;
     }
 }
